Move account data factory to background testing utils
