+++ ui/tests/factories.ts
@@ -1,33 +1,15 @@
 import { Ability } from "@tallyho/tally-background/abilities"
 import { ETHEREUM } from "@tallyho/tally-background/constants"
-import {
-  AccountData,
-  AccountState,
-} from "@tallyho/tally-background/redux-slices/accounts"
+import { AccountState } from "@tallyho/tally-background/redux-slices/accounts"
 import {
   NFTCached,
   NFTCollectionCached,
 } from "@tallyho/tally-background/redux-slices/nfts_update"
+import { createAccountData } from "@tallyho/tally-background/tests/factories"
 import { NormalizedEVMAddress } from "@tallyho/tally-background/types"
 
 export const TEST_ADDRESS =
   "0x208e94d5661a73360d9387d3ca169e5c130090cd" as NormalizedEVMAddress
-
-export const createAccountData = (
-  overrides: Partial<AccountData> = {}
-): AccountData => {
-  return {
-    address: TEST_ADDRESS,
-    network: ETHEREUM,
-    balances: {},
-    ens: {
-      name: "test.crypto",
-    },
-    defaultName: "Test",
-    defaultAvatar: "test.png",
-    ...overrides,
-  }
-}
 
 export const createAccountState = (
   overrides: Partial<AccountState> = {}
